it displays the actual error now~ 	modified:   exmmtest.exe 	modified:   src/lib/16_mm.c 	modified:   src/lib/16_mm.h

+++ src/lib/16_mm.h
@@ -163,9 +163,9 @@ typedef struct
 //==========================================================================
 
 boolean MML_CheckForEMS(void);
-unsigned MML_SetupEMS(mminfo_t *mm);
+byte MML_SetupEMS(mminfo_t *mm);
 void MML_ShutdownEMS(mminfo_t *mm);
-unsigned MM_MapEMS(mminfo_t *mm);
+byte MM_MapEMS(mminfo_t *mm);
 boolean MML_CheckForXMS(mminfo_t *mm);
 void MML_SetupXMS(mminfo_t *mm, mminfotype *mmi);
 void MML_ShutdownXMS(mminfo_t *mm);
